No phi_s for empty RF station
- Since the computation of phi_s, Q_s, and omega_s0 for 'empty' RF station gave a RunTimeWarning, these attributes are now omitted if rf_station.empty is True
- Added some unittests for this
- Also renamed test_general_parameters.py to the appropriate test_ring.py
